Add files via upload
Need to make more experiments but that's about it. I would be nice to have more test problems to show off but that might be a bit difficult. It also might be worth it to make a thing that generates results but does not show them and records data about HV and such so we can get better statistical analysis. That might also take a bit though.